backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "langchain>=0.3.26",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-mongodb>=0.1.4",
    "google-api-python-client>=2.118.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "langchain-tavily>=0.2.3",
    "curl-cffi>=0.11.4",
    "markdownify>=1.1.0",
    "readabilipy>=0.3.0",
    "beautifulsoup4>=4.13.4",
    "markdown>=3.4.0",
    "dotenv>=0.9.9",
    "sse-starlette>=2.3.6",
    "langchain-community>=0.3.26",
    "uvicorn>=0.34.3",
    "langchain-openai>=0.3.24",
    "streamlit>=1.46.0",
    "requests>=2.32.4",
    "aiohttp>=3.12.13",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "msgpack>=1.1.1",
]
